J2ME: Optionally fix sprite clipping by checking render box for camera instead of collision box

+++ src/GbaMonoGame.Rayman3/J2ME/Scene/Box.cs
@@ -2,8 +2,8 @@
 
 public readonly struct Box
 {
-    public sbyte Left { get; init; }
-    public sbyte Top { get; init; }
-    public sbyte Right { get; init; }
-    public sbyte Bottom { get; init; }
+    public int Left { get; init; }
+    public int Top { get; init; }
+    public int Right { get; init; }
+    public int Bottom { get; init; }
 }
